+++ priv/templates/coherence.install/emails/coherence/user_email.ex
@@ -46,9 +46,9 @@ defmodule <%= base %>.Coherence.UserEmail do
 
   defp add_reply_to(mail) do
     case Coherence.Config.email_reply_to do
-      nil     -> mail
-      true    -> reply_to mail, from_email
-      address -> reply_to mail, address
+      nil              -> mail
+      true             -> reply_to mail, from_email
+      address          -> reply_to mail, address
     end
   end
 
@@ -66,9 +66,15 @@ defmodule <%= base %>.Coherence.UserEmail do
   defp from_email do
     case Coherence.Config.email_from do
       nil ->
-        Logger.error ~s|Need to configure :coherence, :email_from, {"Name", "me@example.com"}|
-        ""
-      email -> email
+        Logger.error ~s|Need to configure :coherence, :email_from_name, "Name", and :email_from_email, "me@example.com"|
+        nil
+      {name, email} = email_tuple ->
+        if is_nil(name) or is_nil(email) do
+          Logger.error ~s|Need to configure :coherence, :email_from_name, "Name", and :email_from_email, "me@example.com"|
+          nil
+        else
+          email_tuple
+        end
     end
   end
 end